+++ changelog.md
@@ -98,6 +98,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Add the `--file <file>` option to the `uiua repl` command
   - This runs a file before starting the REPL
 - Improve the supported binding type coverage of [`&ffi`](https://uiua.org/docs/&ffi)
+- Add warnings for when a loop in an array may have a variable signature
 - Various performance improvements
 - Lots of bug and crash fixes
 ### Website